+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- LocalDataPostEvent: add options for, e.g. localized local data attributes
 - LocalData: add optional entity_short_name attribute to local data and local data subscriber configuration to specify the entity the local data
 attribute is defined for. It can be omitted if the local data attribute is defined for all entities or there is only one entity using
 local data attributes.